05-agent-parallel/requirements.txt
mcp
openai>=1.0.0
python-dotenv>=1.0.0
pytz
websockets
